pe_diligence_rag/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Vector Store
faiss-cpu>=1.7.4

# Embeddings (Free, local)
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0

# LLM (Groq - free)
groq>=0.4.0

# Web Server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# WhatsApp
twilio>=8.10.0

# SEC Data
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
